Move refund calculation into the dialog, so that we can set an initial refund value when the dialog is opened. Fixes #18673
